Add plumbing for debug info through dawn_wire.
The debug info payloads are opaque binary data as far as dawn is
concerned. It up to the embedder to encode and decode the payloads.
To use this feature, the embedder must provide a DebugInfoProvider
at WireClient construction time.  The DebugInfoProvider will be
invoked at each serialized command in order to extract debugging
information on a per command basis.  On the service side, the
embedder must provide a DebugErrorHandler at WireServer construction
time.  The DebugErrorHandler is invoked when a service side error
occurs and receives the debug data payload associated with the
command that was being handled at the time that the error occured

This feature can be used, for example, to print client-side stack
traces when an error occurs on the service side.